+++ Tests/UIKitTests/UILabelExtensionsTests.swift
@@ -19,6 +19,16 @@ final class UILabelExtensionsTests: XCTestCase {
         XCTAssertEqual(label.text, "Hello world")
     }
 
+    func testInitWithTextAndStyle() {
+        let style = UIFont.TextStyle.headline
+        let text = "Hello world"
+        let label = UILabel(text: text, style: style)
+
+        let preferredFont = UIFont.preferredFont(forTextStyle: style)
+        XCTAssertEqual(label.text, text)
+        XCTAssertEqual(label.font, preferredFont)
+    }
+
     func testrequiredHeight() {
         let frame = CGRect(x: 0, y: 0, width: 100, height: 100)
         let label = UILabel(frame: frame)